Switched how dev mode is entered
Switching to a "#define" based behaviour means that the system will only be compiled with the code for one mode or the other. Making it smaller and faster by removing the repeated "if" check. In more advanced IDEs than Arduino such as VS Code it will generally show you which set of code is active too.